# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NHCE Campus Connect API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/me/sync:
    post:
      operationId: syncUser
      tags: [users]
      summary: Sync Clerk user into DB (creates or updates)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSync"
      responses:
        "200":
          description: Synced user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: role
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── POSTS ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, clerkId, name, email, role, createdAt]
      properties:
        id:          { type: string }
        clerkId:     { type: string }
        name:        { type: string }
        email:       { type: string }
        role:        { type: string }
        usn:         { type: ["string", "null"] }
        branch:      { type: ["string", "null"] }
        year:        { type: ["integer", "null"] }
        course:      { type: ["string", "null"] }
        avatarUrl:   { type: ["string", "null"] }
        isAdmin:     { type: boolean }
        createdAt:   { type: string }
    UserSync:
      type: object
      required: [clerkId, name, email]
      properties:
        clerkId:   { type: string }
        name:      { type: string }
        email:     { type: string }
        avatarUrl: { type: ["string", "null"] }
    UserUpdate:
      type: object
      properties:
        name:    { type: string }
        usn:     { type: string }
        branch:  { type: string }
        year:    { type: integer }
        course:  { type: string }
        role:    { type: string }
    UserListResponse:
      type: object
      required: [users, total]
      properties:
        users: { type: array, items: { $ref: "#/components/schemas/User" } }
        total: { type: integer }

    # ── POST ────────────────────────────────────────────────────────────────
...